Create router folder
