+++ app/main.py
@@ -1,12 +1,10 @@
-from fastapi import Depends, FastAPI, HTTPException
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from datetime import datetime
 from .features.git import Git
 import os
 import json
-from . import crud, models, schemas
-from .database import SessionLocal, engine
+from .routers import events
 
 app = FastAPI(
     swagger_ui_parameters={
@@ -43,43 +41,4 @@ async def root():
 def health_check():
     return "success"
 
-
-# Dependency
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-@app.post("/event/", response_model=schemas.Event)
-def create_event(event: schemas.EventCreate, db: Session = Depends(get_db)):
-    return crud.create_event(db=db, event=event)
-
-
-@app.get("/events/", response_model=list[schemas.Event])
-def read_events(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    events = crud.get_events(db, skip=skip, limit=limit)
-    return events
-
-
-@app.get("/events/{event_id}", response_model=schemas.Event)
-def read_event(event_id: int, db: Session = Depends(get_db)):
-    db_event = crud.get_event(db, event_id=event_id)
-    if db_event is None:
-        raise HTTPException(status_code=404, detail="Event not found")
-    return db_event
-
-
-# @app.post("/users/{user_id}/items/", response_model=schemas.Item)
-# def create_item_for_user(
-#     user_id: int, item: schemas.ItemCreate, db: Session = Depends(get_db)
-# ):
-#     return crud.create_user_item(db=db, item=item, user_id=user_id)
-
-
-# @app.get("/items/", response_model=list[schemas.Item])
-# def read_items(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-#     items = crud.get_items(db, skip=skip, limit=limit)
-#     return items
+app.include_router(events.router)
